Add `Curator` entity, configuration, and migration; seed admin curator data; update email conflict checks in employer creation.

+++ src/Launchpad/Launchpad.Api/Services/Interfaces/ICurrentUserService.cs
@@ -8,7 +8,7 @@ public interface ICurrentUserService
     /// <summary>
     ///     Profile ID
     /// </summary>
-    public int ProfileId { get; }
+    public long ProfileId { get; }
 
     /// <summary>
     ///     Contact email
